// apps/web/components/submit-button.d.ts
export declare const SubmitButton: ({ isSubmitting, children, className, disabled, onClick, ...props }: {
    isSubmitting?: boolean;
    children: React.ReactNode | string;
    className?: string;
    onClick?: () => void;
    disabled?: boolean;
}) => import("react").JSX.Element;
//# sourceMappingURL=submit-button.d.ts.map